+++ deduce/patterns/name_patterns.py
